font-microservice/requirements.txt
# Font Recognition Microservice — Python Dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Image processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# OCR — text detection and recognition
paddlepaddle==2.6.2
paddleocr==2.9.1

# CLIP — semantic visual embeddings (Phase 4)
# Downloads ViT-B/32 model (~350MB) on first run
sentence-transformers==3.0.1
torch==2.4.0
torchvision==0.19.0

# Font rendering — render font files to images for indexing
fonttools==4.53.1
freetype-py==2.4.0

# Database — store CLIP vectors
psycopg2-binary==2.9.9
pgvector==0.3.2

# Utilities
python-multipart==0.0.9
httpx==0.27.2
python-dotenv==1.0.1
scipy==1.14.0
